fix clang-tidy.....

+++ tasks/vlasova_a_matrix_multiply_ccs/seq/src/ops_seq.cpp
@@ -89,7 +89,7 @@ void ProcessColumnSEQ(const SparseMatrixCCS &at, const SparseMatrixCCS &b, int c
     }
   }
 
-  for (int i = 0; i < static_cast<int>(temp_row.size()); i++) {
+  for (size_t i = 0; i < temp_row.size(); i++) {
     if (row_marker[i] == col_index && std::abs(temp_row[i]) > kEpsilon) {
       res_val.push_back(temp_row[i]);
       res_row_ind.push_back(i);
